requirements.txt
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-extensions==3.1.5
drf-yasg==1.20.0
mysqlclient==2.1.0
drf-nested-routers==0.93.4